profileapp/requirements.txt
aniso8601==7.0.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
bcrypt==4.1.3
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
crispy-bootstrap5==2023.10
cryptography==42.0.5
cssselect2==0.7.0
dateparser==1.2.0
datetime-truncate==1.1.1
defusedxml==0.8.0rc2
dj-database-url==2.2.0
Django==4.2.7
django-allauth==0.63.1
django-bower==5.2.0
django-colorfield==0.11.0
django-crispy-forms==2.1
django-jazzmin==2.6.1
django-jet-reboot==1.3.7
django-memoize==2.3.1
django-multiselectfield==0.1.12
django-nvd3==0.9.7
django-shortcuts==1.6
Faker==25.3.0
feedparser==6.0.11
google-api-python-client==1.4.1
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
html5lib==1.1
httplib2==0.22.0
idna==3.7
jet-bridge-base==1.8.43
Jinja2==3.1.3
lxml==5.2.1
MarkupSafe==2.1.5
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
oscrypto==1.3.0
packaging==24.1
pandas==2.2.2
paramiko==2.8.1
pillow==10.2.0
promise==2.3
prompt-toolkit==2.0.9
psutil==5.9.8
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyHanko==0.24.0
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
PyNaCl==1.5.0
pyotp==2.9.0
pyparsing==3.1.2
pypdf==4.2.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-nvd3==0.14.2
python-slugify==1.1.4
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
regex==2024.5.15
reportlab==4.0.9
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
Rx==1.6.3
sgmllib3k==1.0.0
six==1.16.0
social-auth-core==4.5.4
SQLAlchemy==1.4.52
sqlparse==0.4.4
sshtunnel==0.4.0
svglib==1.5.1
text-unidecode==1.3
tinycss2==1.3.0
tornado==5.1.1
typing_extensions==4.11.0
tzdata==2023.3
tzlocal==5.2
Unidecode==1.3.8
uritemplate==4.1.1
uritools==4.0.2
urllib3==2.2.1
uvicorn==0.30.1
wcwidth==0.2.13
webencodings==0.5.1
whitenoise==6.6.0
xhtml2pdf==0.2.15
